requirements.txt
python-dateutil==2.8.1
python-engineio==3.1.2
python-socketio==3.0.0
eventlet==0.23.0
flask==1.1.2
flask-socketio==3.0.1
numpy
matplotlib
jupyter
pillow
scikit-learn
scikit-image
scipy
h5py
seaborn
pandas
ffmpeg
imageio
tensorflow-cpu
moviepy
opencv-python
requests
PyDot
GraphViz